+++ src/main.cpp
@@ -48,7 +48,7 @@
 namespace
 {
 
-constexpr auto version = std::string_view{"0.10.41"};
+constexpr auto version = std::string_view{"0.10.42"};
 
 struct BootstrapConfigResult final
 {